Create game rps with winner at 5 points
Note the following. Somewhat weird use of variables playersChoice and
playerSelection. Very unelegant code. Wouldn't it be possible
to exchange a few lines of code incessantly repeatead for functions?
Wouldn't it be possible to remove a few lines of code in regards to the
if statements checking if any of the players have reached five points?